linux/kvm_base: fix 32-bit x86 build
gcc -m32 complains about casting a 32-bit pointer to a 64-bit integer:

    kvm_base.c: In function ‘main’:
    kvm_base.c:286:29: error: cast from pointer to integer of different
    size [-Werror=pointer-to-int-cast]
         memreg.userspace_addr = (uint64_t)guest_mempage;
                                 ^

+++ linux/kvm_base.c
@@ -283,7 +283,7 @@ int main(void)
     memreg.flags = 0;
     memreg.guest_phys_addr = 0xfffff000;
     memreg.memory_size = 0x1000;
-    memreg.userspace_addr = (uint64_t)guest_mempage;
+    memreg.userspace_addr = (uint64_t)(uintptr_t)guest_mempage;
     if (ioctl(vmfd, KVM_SET_USER_MEMORY_REGION, &memreg) < 0) {
         perror("ioctl(KVM_SET_USER_MEMORY_REGION, code)");
         return 1;
